requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
librosa>=0.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3
sentence-transformers>=2.2.0
audioread>=3.0.0
soundfile>=0.12.0
tqdm>=4.65.0
